+++ Common/Global/Reverie.GlobalItem.cs
